CHG: Split basic nodes and math nodes
